+++ src/stonks_overwatch/config/bitvavo.py
@@ -25,7 +25,7 @@ class BitvavoConfig(BaseConfig):
         self,
         credentials: Optional[BitvavoCredentials],
         update_frequency_minutes: int = DEFAULT_BITVAVO_UPDATE_FREQUENCY,
-        enabled: bool = True,
+        enabled: bool = False,
         offline_mode: bool = False,
     ) -> None:
         super().__init__(credentials, enabled)
@@ -68,7 +68,7 @@ class BitvavoConfig(BaseConfig):
     @classmethod
     def default(cls) -> "BitvavoConfig":
         try:
-            return cls.from_json_file(cls.CONFIG_PATH)
+            return cls.from_db_with_json_override("bitvavo")
         except Exception:
             cls.logger.debug("Cannot find BITVAVO configuration file. Using default values")
             return BitvavoConfig(