+++ src/stonks_overwatch/models.py
@@ -2,4 +2,5 @@
 from stonks_overwatch.services.brokers.bitvavo.repositories.models import *
 from stonks_overwatch.services.brokers.degiro.repositories.models import *
 from stonks_overwatch.services.brokers.ibkr.repositories.models import *
+from stonks_overwatch.services.brokers.models import *
 from stonks_overwatch.services.brokers.yfinance.repositories.models import *